data_structure::linked_list
Struct ListNode

pub struct ListNode<T> {
    pub val: T,
    pub next: Option<Box<ListNode<T>>>,
}

Fields
val: Tnext: Option<Box<ListNode<T>>>
Implementations

impl<T> ListNode<T>

pub fn new(val: T) -> Box<Self>

Constructs a new instance of ListNode<T> with the provided val: T, returning Box<ListNode<T>>

Example

let linked_list = ListNode::new(0);
assert_eq!(linked_list, Box::new( ListNode{ val: 0, next: None } ));

Variable linked_list will have a value of 0 and next pointing to None:

pub fn from_vec(l: Vec<T>) -> Box<Self>
where T: Clone,

Constructs a new instance of ListNode<T> with a Vec<T>, returning Box<ListNode<T>>

Example

assert_eq!(
    ListNode::from_vec(vec![1, 2, 3]),
    Box::new(ListNode {
        val: 1,
        next: Some(Box::new(ListNode {
            val: 2,
            next: Some(Box::new(ListNode {
                val: 3,
                next: None,
            }))
        }))
    })
);

pub fn print(&self)
where T: Display,

Prints the provided ListNode<T>

Example

let list = ListNode::from_vec(vec![1, 2, 3]);
list.print(); //Output = 1 -> 2 -> 3 -> None

pub fn push(&mut self, val: T)
where T: Clone,

Pushes an instance of ListNode<T> to the front of the list

Example

let mut list = ListNode::from_vec(vec![2, 3]);
list.push(1);
assert_eq!(list, ListNode::from_vec(vec![1, 2, 3]));

pub fn push_back(&mut self, val: T)
where T: Copy,

Pushes an instance of ListNode<T> to the back of the list

Example

let mut list = ListNode::from_vec(vec![1, 2]);
list.push_back(3);
assert_eq!(list, ListNode::from_vec(vec![1, 2, 3]));

pub fn delete(&mut self, val: T) -> Result<(), &'static str>
where T: PartialEq,

Deletes the node that equals to the given val: T

The function will delete the first node in the sequence iterating from head, not by index.

If the element is not found, the program would panic, yielding Node not found.

Example

let mut list = ListNode::from_vec(vec![1, 2, 2, 3]);
list.delete(2)?;
assert_eq!(list, ListNode::from_vec(vec![1, 2, 3]));

pub fn find(&mut self, val: T) -> Result<&Box<Self>, &'static str>
where T: PartialEq,

Finds the node with the give val: T and returns a reference of element.

This function does not return a new copy of the linked-list, it returns a reference to a node already on the list.

Example

 let mut list = ListNode::from_vec(vec![1, 2, 3]);
 let found_node = list.find(2)?;
 assert_eq!(found_node.val, 2);
 // 1 (list's head) -> 2 (found_node's head) -> 3;
 // find() does not copy the original list
 // so the two pointers could potentially
 // be the same element
 let mut list2 = ListNode::from_vec(vec![1, 2, 3]);
 {
    let found_node2 = list2.find(1)?;
    assert_eq!(found_node2.val, 1);
 }
 // compares the pointers
 assert_eq!(
    format!("{:p}", list2.as_ref()),
    format!("{:p}", list2.find(1)?.as_ref())
);
 // 1 (list2's head and found_node2's head) -> 2 -> 3
 // you could print the list as well, as the
 // print function prints the pointer address too

pub fn len(&mut self) -> i32

Return the length of the given ListNode<T> as an i32

Example

let mut list = ListNode::from_vec(vec![1, 2, 3]);
assert_eq!(list.len(), 3);

pub fn reverse(&mut self)
where T: Copy,

Reverses the given ListNode<T>

Example

let mut list = ListNode::from_vec(vec![1, 2, 3]);
let mut rev_list = ListNode::from_vec(vec![3, 2, 1]);
rev_list.reverse();
assert_eq!(list, rev_list);

pub fn copy(&self) -> Box<ListNode<T>>
where T: Clone,

Deep copies the given ListNode<T>

Example

let list = ListNode::from_vec(vec![1, 2, 3]);
let copied = list.copy();
assert_eq!(list, copied);
// compares the pointers
assert_ne!(
    (format!(
        "{:p}{:p}{:p}",
        list,
        list.next.as_ref().unwrap(),
        list.next.as_ref().unwrap().next.as_ref().unwrap())),
    (format!(
        "{:p}{:p}{:p}",
        copied,
        copied.next.as_ref().unwrap(),
        copied.next.as_ref().unwrap().next.as_ref().unwrap())),
);

pub fn insert(&mut self, index: usize, val: T) -> Result<(), &'static str>

Inserts a new node of ListNode<T> with the val: T and in position index: usize

if the index is out of range, the function would return Err

Example

 let mut list = ListNode::from_vec(vec![1, 3]);
 list.insert(1, 2)?;
 assert_eq!(list, ListNode::from_vec(vec![1, 2, 3]));

pub fn pop(&mut self) -> Option<T>

Pops the element on the front of the list

Example

let mut list = ListNode::from_vec(vec![1, 1, 2, 3]);
list.pop()?;
assert_eq!(list, ListNode::from_vec(vec![1, 2, 3]));

pub fn contains(&self, val: T) -> bool
where T: PartialEq,

Checks if whether the ListNode<T> contains the given val: T element

Example

let list = ListNode::from_vec(vec![1, 2, 3]);
assert_eq!(list.contains(2), true);

pub fn merge( l1: Option<Box<ListNode<i32>>>, l2: Option<Box<ListNode<i32>>>, ) -> Option<Box<ListNode<i32>>>
where T: PartialOrd,

Merges two sorted ListNode<T> while keeping the order

Example

let temp1 = ListNode::from_vec(vec![1, 3]);
let temp2 = ListNode::from_vec(vec![2, 4, 5]);
let list = ListNode::<i32>::merge(Some(temp1), Some(temp2))?;
assert_eq!(list, ListNode::from_vec(vec![1, 2, 3, 4, 5]));

pub fn sort(head: Option<Box<ListNode<i32>>>) -> Option<Box<ListNode<i32>>>

Sorts the ListNode<T> through merge sort

Example

let list = ListNode::<i32>::sort(Some(ListNode::from_vec(vec![5, 2, 3, 1, 4])))?;
assert_eq!(list, ListNode::from_vec(vec![1, 2, 3, 4, 5]));